Renamed SDL_ttf.h to SDL_ttf_fontfallback.h.
