+++ src/Game.cpp
@@ -41,7 +41,7 @@
 #include <assert.h>
 #include <stdio.h>
 #include <string.h>
-#include <SDL_ttf.h>
+#include <SDL_ttf_fontfallback.h>
 
 #include "libs/tinyformat/tinyformat.h"
