+++ src/LevelPlaySelect.cpp
@@ -39,7 +39,7 @@
 #include <sstream>
 #include <iostream>
 
-#include <SDL_ttf.h>
+#include <SDL_ttf_fontfallback.h>
 
 #include "libs/tinyformat/tinyformat.h"
